Refine cabinet game modals
